moved nav function bulk to commands module
